// Reclutamiento/ClientApp/src/app/shared/enums/data-type.enum.ts
export enum DataType {
  Date = 'date',
  DateDash = 'dateDash',
  DateFull = 'datefull',
  Currency = 'currency',
  String = 'string',
  Link = 'link',
  Number = 'number',
  DateFullTimeMedium = 'datefulltimemedium',
  DateFullTimeFull = 'datefulltimefull',
  PopUp = 'popup',
  PopUpDateTime = 'popupdatetime'
}